PlugStatus v0.1.0 PlugStatus

A plug for responding to status requests, returns performance metrics.

This plug responds with a successful status to GET requests at a specific path so that clients can get status information.

The response that this plug sends is a 200 response with a json body containing status information, the path that responds to this request is /__status, but it can be configured.

Note that this plug halts the connection. This is done so that it can be plugged near the top of a plug pipeline and catch requests early so that subsequent plugs don’t have the chance to tamper with the connection. Read more about halting connections in the docs for Plug.Builder.

Information returned

The following information below is returned.

memory - all information from Erlang VM call memory/0

statistics - These items from Erlang VM call statistics/1: :active_tasks, :context_switches, :garbage_collection, :io, :run_queue, :run_queue_lengths, :scheduler_wall_time, :total_active_tasks, :total_run_queue_lengths

system_info - These items from Erlang VM call system_info/1: :atom_count, :atom_limit, :otp_release, :port_count, :port_limit, :process_count, :process_limit, :schedulers, :schedulers_online, :thread_pool_size, :version

Options

The following options can be used when calling plug PlugStatus.

:path - a string specifying the path on which PlugStatus will respond to status requests

Examples

defmodule MyServer do
  use Plug.Builder
  plug PlugStatus

  # ... rest of the pipeline
end

Using a custom path:

defmodule MyServer do
  use Plug.Builder
  plug PlugStatus, path: "/status"

  # ... rest of the pipeline
end

Functions

call(conn, opts)

Callback implementation for Plug.call/2.

init(opts)

Callback implementation for Plug.init/1.